ep1/requirements.txt
# python
wheel
neovim
ipython
ipykernel
python-dotenv

# code
black[jupyter]

# ops
dagster
dagit

# data
fsspec
adlfs
dask
distributed
deltalake
duckdb --pre
polars[all]
ibis-framework[polars,duckdb]

# ml
xgboost

# viz
matplotlib
seaborn
bokeh
plotly

# mlops
mlflow
